# expression: implement vectorized evaluation for JSONInsert JSONSet JSONReplace