<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/community/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->
PCP #12104 

### What problem does this PR solve? <!--add issue link with summary if exists-->
implement vectorized evaluation for JSONInsert, JSONSet, JSONReplace by implementing `vecJSONModify`. #12104 

### What is changed and how it works?
```
BenchmarkVectorizedBuiltinJSONFunc/builtinJSONSetSig-VecBuiltinFunc-4                        452           2650401 ns/op         1118522 B/op      11222 allocs/op
BenchmarkVectorizedBuiltinJSONFunc/builtinJSONSetSig-NonVecBuiltinFunc-4                     441           2691375 ns/op         1118416 B/op      11220 allocs/op
BenchmarkVectorizedBuiltinJSONFunc/builtinJSONReplaceSig-VecBuiltinFunc-4                    960           1204402 ns/op          450550 B/op       5062 allocs/op
BenchmarkVectorizedBuiltinJSONFunc/builtinJSONReplaceSig-NonVecBuiltinFunc-4                 963           1261635 ns/op          450602 B/op       5060 allocs/op
BenchmarkVectorizedBuiltinJSONFunc/builtinJSONInsertSig-VecBuiltinFunc-4                     402           2951319 ns/op         1377495 B/op      13351 allocs/op
BenchmarkVectorizedBuiltinJSONFunc/builtinJSONInsertSig-NonVecBuiltinFunc-4                  397           3672038 ns/op         1377460 B/op      13349 allocs/op
```

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test